@@ -1,18 +1,18 @@
 {
-  "name": "vaersaagod/bunny",
-  "description": "Bunny CDN integration for Craft CMS.",
+  "name": "vaersaagod/bunnymate",
+  "description": "Keeping your bunny finely-tuned for a hoppy life, mate!",
   "type": "craft-plugin",
-  "version": "2.0.0",
+  "version": "1.0.0",
   "keywords": [
     "craft",
     "cms",
     "craftcms",
     "craft-plugin",
-    "bunny"
+    "bunnymate"
   ],
   "support": {
-    "docs": "https://github.com/vaersaagod/bunny/blob/master/README.md",
-    "issues": "https://github.com/vaersaagod/bunny/issues"
+    "docs": "https://github.com/vaersaagod/bunnymate/blob/main/README.md",
+    "issues": "https://github.com/vaersaagod/bunnymate/issues"
   },
   "license": "MIT",
   "authors": [
@@ -22,21 +22,21 @@
     }
   ],
   "require": {
-    "php": "^8.0",
+    "php": "^8.1",
     "craftcms/cms": "^4.0.0"
   },
   "autoload": {
     "psr-4": {
-      "vaersaagod\\bunny\\": "src/"
+      "vaersaagod\\bunnymate\\": "src/"
     }
   },
   "extra": {
-    "name": "Bunny",
-    "handle": "bunny",
+    "name": "BunnyMate",
+    "handle": "_bunnymate",
     "developer": "Værsågod",
     "developerUrl": "https://vaersaagod.no",
-    "documentationUrl": "https://github.com/vaersaagod/bunny/blob/master/README.md",
-    "changelogUrl": "https://raw.githubusercontent.com/vaersaagod/bunny/master/CHANGELOG.md",
-    "class": "vaersaagod\\bunny\\Bunny"
+    "documentationUrl": "https://github.com/vaersaagod/bunnymate/blob/main/README.md",
+    "changelogUrl": "https://raw.githubusercontent.com/vaersaagod/bunny/main/CHANGELOG.md",
+    "class": "vaersaagod\\bunnymate\\BunnyMate"
   }
 }
